Extract Todo API executable to a command package
