+++ Dockerfile
@@ -1,16 +1,23 @@
 # Build stage
 FROM golang:alpine as build-stage
+
 ENV GO111MODULE=on \
     CGO_ENABLED=0 \
     GOOS=linux \
     GOARCH=amd64
+
+RUN apk update && apk --no-cache add make
+
 WORKDIR /build 
-COPY . .
+COPY go.mod .
+COPY go.sum .
 RUN go mod download
-RUN go build -o main .
+COPY . .
+
+RUN make build
 
 # Run stage
 FROM scratch
-COPY --from=build-stage /build/main /
+COPY --from=build-stage /build/todoapi /
 EXPOSE 3000
-CMD ["/main"]
+CMD ["/todoapi"]
