+++ todo_test.go
@@ -1,4 +1,4 @@
-package main
+package todo
 
 import (
 	"io/ioutil"
